refactor flow manager into more generic flow runner to be reusable in control panel
