+++ Photon-Tinker/Mesh/MeshSetupFlowManager.swift
@@ -6,10 +6,9 @@
 import Foundation
 
 
-class MeshSetupFlowManager: NSObject, MeshSetupBluetoothConnectionManagerDelegate, MeshSetupStepDelegate, MeshSetupFlowManagerDelegateResponseConsumer {
+class MeshSetupFlowManager : MeshSetupFlowRunner {
 
-
-    private let preflow:[MeshSetupStep] = [
+    fileprivate let preflow:[MeshSetupStep] = [
         StepGetTargetDeviceInfo(),
         StepConnectToTargetDevice(),
         StepEnsureCorrectEthernetFeatureStatus(),
@@ -21,7 +20,7 @@ class MeshSetupFlowManager: NSObject, MeshSetupBluetoothConnectionManagerDelegat
         StepCheckTargetDeviceHasNetworkInterfaces(),
     ]
 
-    private let joinerFlow: [MeshSetupStep] = [
+    fileprivate let joinerFlow: [MeshSetupStep] = [
         StepShowInfo(),
         StepGetUserNetworkSelection(),
         StepGetCommissionerDeviceInfo(),
@@ -37,13 +36,13 @@ class MeshSetupFlowManager: NSObject, MeshSetupBluetoothConnectionManagerDelegat
     ]
 
     //runs before ethernet/wifi/cellular flows
-    private let internetConnectedPreflow: [MeshSetupStep] = [
+    fileprivate let internetConnectedPreflow: [MeshSetupStep] = [
         StepOfferSetupStandAloneOrWithNetwork(),
         StepOfferSelectOrCreateNetwork()
     ]
 
 
-    private let ethernetFlow: [MeshSetupStep] = [
+    fileprivate let ethernetFlow: [MeshSetupStep] = [
         StepShowPricingImpact(),
         StepShowInfo(),
         StepEnsureHasInternetAccess(),
@@ -51,7 +50,7 @@ class MeshSetupFlowManager: NSObject, MeshSetupBluetoothConnectionManagerDelegat
         StepPublishDeviceSetupDoneEvent()
     ]
 
-    private let wifiFlow: [MeshSetupStep] = [
+    fileprivate let wifiFlow: [MeshSetupStep] = [
         StepShowPricingImpact(),
         StepShowInfo(),
         StepGetUserWifiNetworkSelection(),
@@ -61,7 +60,7 @@ class MeshSetupFlowManager: NSObject, MeshSetupBluetoothConnectionManagerDelegat
         StepPublishDeviceSetupDoneEvent()
     ]
 
-    private let cellularFlow: [MeshSetupStep] = [
+    fileprivate let cellularFlow: [MeshSetupStep] = [
         StepShowPricingImpact(),
         StepShowInfo(),
         StepEnsureHasInternetAccess(),
@@ -70,7 +69,7 @@ class MeshSetupFlowManager: NSObject, MeshSetupBluetoothConnectionManagerDelegat
     ]
 
     //runs post ethernet/wifi/cellular flows
-    private let networkCreatorPostflow: [MeshSetupStep] = [
+    fileprivate let networkCreatorPostflow: [MeshSetupStep] = [
         StepGetNewDeviceName(),
         StepGetNewNetworkName(),
         StepGetNewNetworkPassword(),
@@ -81,52 +80,11 @@ class MeshSetupFlowManager: NSObject, MeshSetupBluetoothConnectionManagerDelegat
     ]
 
     //runs post ethernet/wifi/cellular flows
-    private let standalonePostflow: [MeshSetupStep] = [
+    fileprivate let standalonePostflow: [MeshSetupStep] = [
         StepGetNewDeviceName(),
         StepOfferToAddOneMoreDevice()
     ]
 
-
-    private (set) public var context: MeshSetupContext
-
-    private var currentFlow: [MeshSetupStep]!
-    private var currentStepIdx: Int = 0
-    private var currentStep: MeshSetupStep {
-        return currentFlow[currentStepIdx]
-    }
-
-
-    init(delegate: MeshSetupFlowManagerDelegate) {
-        self.context = MeshSetupContext()
-
-        super.init()
-
-        context.delegate = delegate
-        context.stepDelegate = self
-        context.bluetoothManager = MeshSetupBluetoothConnectionManager(delegate: self)
-    }
-
-    internal func log(_ message: String) {
-        ParticleLogger.logInfo("MeshSetupFlow", format: message, withParameters: getVaList([]))
-    }
-
-    internal func fail(_ sender: MeshSetupStep, withReason reason: MeshSetupFlowError, severity: MeshSetupErrorSeverity, nsError: Error?) {
-        self.fail(withReason: reason, severity: severity, nsError: nsError)
-    }
-
-    internal func fail(withReason reason: MeshSetupFlowError, severity: MeshSetupErrorSeverity = .Error, nsError: Error? = nil) {
-        if context.canceled == false {
-            if (severity == .Fatal) {
-                self.cancelSetup()
-            }
-
-            self.log("error: \(reason.description), nsError: \(nsError?.localizedDescription as Optional)")
-            context.delegate.meshSetupError(self.currentStep, error: reason, severity: severity, nsError: nsError)
-        }
-    }
-
-
-    //MARK: public interface
     //entry to the flow
     func startSetup() {
         currentFlow = preflow
@@ -135,44 +93,9 @@ class MeshSetupFlowManager: NSObject, MeshSetupBluetoothConnectionManagerDelegat
         self.runCurrentStep()
     }
 
-    func pauseSetup() {
-        context.paused = true
-    }
-
-    func continueSetup() {
-        if (context.paused) {
-            context.paused = false
-            self.runCurrentStep()
-        }
-    }
-
-    func cancelSetup() {
-        context.canceled = true
-
-        context.bluetoothManager.stopScan()
-        context.bluetoothManager.dropAllConnections()
-    }
-
-    private func finishSetup() {
-        context.canceled = true
-
-        context.bluetoothManager.stopScan()
-        context.bluetoothManager.dropAllConnections()
-    }
-
-    internal func rewindTo(_ sender: MeshSetupStep, step: MeshSetupStep.Type) -> MeshSetupStep {
-        if let error = self.rewindTo(step: step) {
-            fatalError("flow tried to perform illegal back")
-        } else {
-            return self.currentStep
-        }
-    }
-
-
     //this is for internal use only, because it requires a lot of internal knowledge to use and is nearly impossible to expose to external developers
-    internal func rewindTo(step: MeshSetupStep.Type) -> MeshSetupFlowError? {
-
-        currentStep.rewindFrom()
+    override internal func rewindTo(step: MeshSetupStep.Type) -> MeshSetupFlowError? {
+        currentStep!.rewindFrom()
 
         if (currentStepIdx == 0) {
             //if we are backing from one of these flows, we need to switch the flow type.
@@ -192,7 +115,7 @@ class MeshSetupFlowManager: NSObject, MeshSetupBluetoothConnectionManagerDelegat
 
                 self.currentStepIdx = i
                 self.log("returning to step: \(self.currentStepIdx)")
-                self.currentStep.rewindTo(context: self.context)
+                self.currentStep!.rewindTo(context: self.context)
                 self.runCurrentStep()
 
                 return nil
@@ -202,32 +125,7 @@ class MeshSetupFlowManager: NSObject, MeshSetupBluetoothConnectionManagerDelegat
         return .IllegalOperation
     }
 
-    func retryLastAction() {
-        self.log("Retrying action: \(self.currentStep)")
-        self.currentStep.retry()
-    }
-
-    //MARK: Flow control
-    private func runCurrentStep() {
-        if (context.canceled) {
-            return
-        }
-
-        //if we reached the end of current flow
-        if (currentStepIdx == currentFlow.count) {
-            self.switchFlow()
-        }
-
-
-        log("stepComplete\n\n" +
-                "--------------------------------------------------------------------------------------------\n" +
-                "currentStepIdx = \(currentStepIdx), currentStep = \(currentStep)")
-
-        self.currentStep.reset()
-        self.currentStep.run(context: self.context)
-    }
-
-    private func switchFlow() {
+    override internal func switchFlow() {
         log("stepComplete\n\n" +
                 "--------------------------------------------------------------------------------------------\n" +
                 "Switching flow!!!")
@@ -265,7 +163,7 @@ class MeshSetupFlowManager: NSObject, MeshSetupBluetoothConnectionManagerDelegat
                 self.currentFlow = networkCreatorPostflow
                 log("setting creatorSubflow flow")
             } else {
-               self.currentFlow = standalonePostflow
+                self.currentFlow = standalonePostflow
                 log("setting standaloneSubflow flow")
             }
         } else {
@@ -275,192 +173,10 @@ class MeshSetupFlowManager: NSObject, MeshSetupBluetoothConnectionManagerDelegat
         self.currentStepIdx = 0
     }
 
-    internal func stepCompleted(_ sender: MeshSetupStep) {
-        if (context.canceled) {
-            return
-        }
-
-        if type(of: sender) != type(of: currentStep) {
-            self.log("Flow order is broken :(. Current command: \(self.currentStep), Parameter command: \(sender)")
-            self.log("Stack:\n\(Thread.callStackSymbols.joined(separator: "\n"))")
-            self.fail(withReason: .CriticalFlowError, severity: .Fatal)
-        }
-
-        self.currentStepIdx += 1
-
-        if (context.paused) {
-            return
-        }
-
-        self.runCurrentStep()
-    }
 
 
-
-
-    //MARK: BluetoothConnectionManagerDelegate
-    internal func bluetoothConnectionManagerStateChanged(sender: MeshSetupBluetoothConnectionManager, state: MeshSetupBluetoothConnectionManagerState) {
-        if (context.canceled) {
-            return
-        }
-
-        self.log("bluetoothConnectionManagerStateChanged = \(state)")
-        if (context.bluetoothManager.state == .Ready) {
-            context.bluetoothReady = true
-        } else if (context.bluetoothManager.state == .Disabled) {
-            context.bluetoothReady = false
-
-            //if we are waiting for the reply = trigger timeout
-            if let targetDeviceTransceiver = context.targetDevice.transceiver {
-                targetDeviceTransceiver.triggerTimeout()
-            }
-
-            //if we are waiting for the reply = trigger timeout
-            if let commissionerDeviceTransceiver = context.commissionerDevice?.transceiver {
-                commissionerDeviceTransceiver.triggerTimeout()
-            }
-        }
-    }
-
-    internal func bluetoothConnectionManagerError(sender: MeshSetupBluetoothConnectionManager, error: BluetoothConnectionManagerError, severity: MeshSetupErrorSeverity) {
-        if (context.canceled) {
-            return
-        }
-
-        log("bluetoothConnectionManagerError = \(error), severity = \(severity)")
-        if (!currentStep.handleBluetoothConnectionManagerError(error)) {
-            self.fail(withReason: .BluetoothError, severity: .Fatal)
-        }
-    }
-
-    internal func bluetoothConnectionManagerConnectionCreated(sender: MeshSetupBluetoothConnectionManager, connection: MeshSetupBluetoothConnection) {
-        if (context.canceled) {
-            return
-        }
-
-        log("bluetoothConnectionManagerConnectionCreated = \(connection)")
-
-        if (!currentStep.handleBluetoothConnectionManagerConnectionCreated(connection)) {
-            self.fail(withReason: .BluetoothError, severity: .Fatal)
-        }
-    }
-
-    internal func bluetoothConnectionManagerConnectionBecameReady(sender: MeshSetupBluetoothConnectionManager, connection: MeshSetupBluetoothConnection) {
-        if (context.canceled) {
-            return
-        }
-
-        log("bluetoothConnectionManagerConnectionBecameReady = \(connection)")
-
-        if (!currentStep.handleBluetoothConnectionManagerConnectionBecameReady(connection)) {
-            self.fail(withReason: .BluetoothError, severity: .Fatal)
-        }
-    }
-
-    internal func bluetoothConnectionManagerConnectionDropped(sender: MeshSetupBluetoothConnectionManager, connection: MeshSetupBluetoothConnection) {
-        if (context.canceled) {
-            return
-        }
-
-        log("bluetoothConnectionManagerConnectionDropped = \(connection)")
-
-        if (connection == context.targetDevice.transceiver?.connection || connection == context.commissionerDevice?.transceiver?.connection) {
-            if (!currentStep.handleBluetoothConnectionManagerConnectionDropped(connection)) {
-                self.fail(withReason: .BluetoothConnectionDropped, severity: .Fatal)
-            }
-        }
-        //if some other connection was dropped - we dont care
-    }
-
-
-
-    func setTargetDeviceInfo(dataMatrix: MeshSetupDataMatrix, useEthernet: Bool) -> MeshSetupFlowError? {
-        guard type(of: currentStep) == StepGetTargetDeviceInfo.self else {
-            return .IllegalOperation
-        }
-
-        return (currentStep as! StepGetTargetDeviceInfo).setTargetDeviceInfo(dataMatrix: dataMatrix, useEthernet: useEthernet)
-    }
-
-    func setTargetPerformFirmwareUpdate(update: Bool) -> MeshSetupFlowError? {
-        guard type(of: currentStep) == StepEnsureLatestFirmware.self else {
-            return .IllegalOperation
-        }
-
-        return (currentStep as! StepEnsureLatestFirmware).setTargetPerformFirmwareUpdate(update: update)
-    }
-
-    func setTargetDeviceLeaveNetwork(leave: Bool) -> MeshSetupFlowError? {
-        guard type(of: currentStep) == StepEnsureTargetDeviceIsNotOnMeshNetwork.self else {
-            return .IllegalOperation
-        }
-
-        return (currentStep as! StepEnsureTargetDeviceIsNotOnMeshNetwork).setTargetDeviceLeaveNetwork(leave: leave)
-    }
-
-
-    func setSelectStandAloneOrMeshSetup(meshSetup: Bool) -> MeshSetupFlowError? {
-        guard type(of: currentStep) == StepOfferSetupStandAloneOrWithNetwork.self else {
-            return .IllegalOperation
-        }
-
-        return (currentStep as! StepOfferSetupStandAloneOrWithNetwork).setSelectStandAloneOrMeshSetup(meshSetup: meshSetup)
-    }
-
-    func setOptionalSelectedNetwork(selectedNetworkExtPanID: String?) -> MeshSetupFlowError? {
-        guard type(of: currentStep) == StepOfferSelectOrCreateNetwork.self else {
-            return .IllegalOperation
-        }
-
-        return (currentStep as! StepOfferSelectOrCreateNetwork).setOptionalSelectedNetwork(selectedNetworkExtPanID: selectedNetworkExtPanID)
-    }
-
-    func rescanNetworks() -> MeshSetupFlowError? {
-        //only allow to rescan if current step asks for it and transceiver is free to be used
-        guard let isBusy = context.targetDevice.transceiver?.isBusy, isBusy == false else {
-            return .IllegalOperation
-        }
-
-        if (type(of: currentStep) == StepOfferSelectOrCreateNetwork.self) {
-            (currentStep as! StepOfferSelectOrCreateNetwork).scanNetworks()
-        } else if (type(of: currentStep) == StepGetUserNetworkSelection.self) {
-            (currentStep as! StepGetUserNetworkSelection).scanNetworks()
-        } else if (type(of: currentStep) == StepGetUserWifiNetworkSelection.self) {
-                (currentStep as! StepGetUserWifiNetworkSelection).scanWifiNetworks()
-        } else {
-            return .IllegalOperation
-        }
-
-        return nil
-    }
-
-    func setPricingImpactDone() -> MeshSetupFlowError? {
-        guard type(of: currentStep) == StepShowPricingImpact.self else {
-            return .IllegalOperation
-        }
-
-        return (currentStep as! StepShowPricingImpact).setPricingImpactDone()
-    }
-
-    func setInfoDone() -> MeshSetupFlowError? {
-        guard type(of: currentStep) == StepShowInfo.self else {
-            return .IllegalOperation
-        }
-
-        return (currentStep as! StepShowInfo).setInfoDone()
-    }
-
-    func setDeviceName(name: String, onComplete:@escaping (MeshSetupFlowError?) -> ()) {
-        guard type(of: currentStep) == StepGetNewDeviceName.self else {
-            onComplete(.IllegalOperation)
-            return
-        }
-
-        (currentStep as! StepGetNewDeviceName).setDeviceName(name: name, onComplete: onComplete)
-    }
-
-    func setAddOneMoreDevice(addOneMoreDevice: Bool) -> MeshSetupFlowError? {
-        guard type(of: currentStep) == StepOfferToAddOneMoreDevice.self else {
+    override func setAddOneMoreDevice(addOneMoreDevice: Bool) -> MeshSetupFlowError? {
+        guard type(of: currentStep!) == StepOfferToAddOneMoreDevice.self else {
             return .IllegalOperation
         }
 
@@ -474,65 +190,5 @@ class MeshSetupFlowManager: NSObject, MeshSetupBluetoothConnectionManagerDelegat
 
         return nil
     }
-
-
-    func setNewNetworkName(name: String) -> MeshSetupFlowError? {
-        guard type(of: currentStep) == StepGetNewNetworkName.self else {
-            return .IllegalOperation
-        }
-
-        return (currentStep as! StepGetNewNetworkName).setNewNetworkName(name: name)
-    }
-
-
-    func setNewNetworkPassword(password: String) -> MeshSetupFlowError? {
-        guard type(of: currentStep) == StepGetNewNetworkPassword.self else {
-            return .IllegalOperation
-        }
-
-        return (currentStep as! StepGetNewNetworkPassword).setNewNetworkPassword(password: password)
-    }
-
-    func setSelectedWifiNetworkPassword(_ password: String, onComplete:@escaping (MeshSetupFlowError?) -> ()) {
-        guard type(of: currentStep) == StepEnsureCorrectSelectedWifiNetworkPassword.self else {
-            onComplete(.IllegalOperation)
-            return
-        }
-
-        (currentStep as! StepEnsureCorrectSelectedWifiNetworkPassword).setSelectedWifiNetworkPassword(password, onComplete: onComplete)
-    }
-
-    func setSelectedWifiNetwork(selectedNetwork: MeshSetupNewWifiNetworkInfo) -> MeshSetupFlowError? {
-        guard type(of: currentStep) == StepGetUserWifiNetworkSelection.self else {
-            return .IllegalOperation
-        }
-
-        return (currentStep as! StepGetUserWifiNetworkSelection).setSelectedWifiNetwork(selectedNetwork: selectedNetwork)
-    }
-
-    func setSelectedNetwork(selectedNetworkExtPanID: String) -> MeshSetupFlowError? {
-        guard type(of: currentStep) == StepGetUserNetworkSelection.self else {
-            return .IllegalOperation
-        }
-
-        return (currentStep as! StepGetUserNetworkSelection).setSelectedNetwork(selectedNetworkExtPanID: selectedNetworkExtPanID)
-    }
-
-    func setCommissionerDeviceInfo(dataMatrix: MeshSetupDataMatrix) -> MeshSetupFlowError? {
-        guard type(of: currentStep) == StepGetCommissionerDeviceInfo.self else {
-            return .IllegalOperation
-        }
-
-        return (currentStep as! StepGetCommissionerDeviceInfo).setCommissionerDeviceInfo(dataMatrix: dataMatrix)
-
-    }
-
-    func setSelectedNetworkPassword(_ password: String, onComplete:@escaping (MeshSetupFlowError?) -> ()) {
-        guard type(of: currentStep) == StepEnsureCorrectSelectedNetworkPassword.self else {
-            onComplete(.IllegalOperation)
-            return
-        }
-
-        (currentStep as! StepEnsureCorrectSelectedNetworkPassword).setSelectedNetworkPassword(password, onComplete: onComplete)
-    }
 }
+
