+++ Photon-Tinker/Mesh/MeshStrings.swift
@@ -54,10 +54,21 @@ extension String {
 }
 
 
+
 class MeshSetupStrings {
     static private let randomNames = ["aardvark", "bacon", "badger", "banjo", "bobcat", "boomer", "captain", "chicken", "cowboy", "maker", "splendid", "sparkling", "dentist", "doctor", "green", "easter", "ferret", "gerbil", "hacker", "hamster", "wizard", "hobbit", "hoosier", "hunter", "jester", "jetpack", "kitty", "laser", "lawyer", "mighty", "monkey", "morphing", "mutant", "narwhal", "ninja", "normal", "penguin", "pirate", "pizza", "plumber", "power", "puppy", "ranger", "raptor", "robot", "scraper", "burrito", "station", "tasty", "trochee", "turkey", "turtle", "vampire", "wombat", "zombie"]
     static func getRandomDeviceName() -> String {
         return "\(MeshSetupStrings.randomNames.randomElement()!)_\(MeshSetupStrings.randomNames.randomElement()!)"
+    }
+
+
+    struct ControlPanel {
+        struct Root {
+            static let Title = "MeshSetup.ControlPanel.Title".meshLocalized()
+
+            static let WiFiTitle = "MeshSetup.ControlPanel.WifiTitle".meshLocalized()
+            static let MeshTitle = "MeshSetup.ControlPanel.MeshTitle".meshLocalized()
+        }
     }
 
     struct Prompt {
