+++ Photon-Tinker/Mesh/StepGetTargetInfo.swift
@@ -20,7 +20,7 @@ class StepGetTargetDeviceInfo: MeshSetupStep {
             return nil
         }
 
-        context.targetDevice = MeshDevice()
+        context.targetDevice = MeshSetupDevice()
 
         self.resetFlowFlags()
 
@@ -31,7 +31,7 @@ class StepGetTargetDeviceInfo: MeshSetupStep {
         context.targetDevice.credentials = MeshSetupPeripheralCredentials(name: self.context!.targetDevice.type!.bluetoothNamePrefix + "-" + dataMatrix.serialNumber.suffix(6), mobileSecret: dataMatrix.mobileSecret)
 
         if (context.targetDevice.credentials?.name == context.commissionerDevice?.credentials?.name) {
-            context.targetDevice = MeshDevice()
+            context.targetDevice = MeshSetupDevice()
             return .ThisDeviceIsACommissioner
         }
 
@@ -43,7 +43,7 @@ class StepGetTargetDeviceInfo: MeshSetupStep {
     override func rewindTo(context: MeshSetupContext) {
         super.rewindTo(context: context)
 
-        context.targetDevice = MeshDevice()
+        context.targetDevice = MeshSetupDevice()
     }
 
     func resetFlowFlags() {
